+++ insta/filter/user_names/Dockerfile
@@ -4,7 +4,7 @@ COPY go.mod .
 COPY go.sum .
 RUN go mod download
 COPY . .
-RUN CGO_ENABLED=0 GOOS=linux go build -installsuffix cgo -o kafka_changestream kafka/changestream/insta_user_names-filter/main.go
+RUN CGO_ENABLED=0 GOOS=linux go build -installsuffix cgo -o kafka_changestream insta/filter/user_names/main.go
 
 FROM alpine
 RUN apk --no-cache add ca-certificates
